backend/requirements.txt
sanic==23.6.0
pymongo==4.4.1
bcrypt==4.0.1
PyJWT==2.8.0
sanic-motor
sanic-cors